# transactions-api/src/swagger.yaml
openapi: 3.0.3
info:
  title: Transactions Service REST API
  version: v1
servers:
  - description: 'Local'
    url: 'http://localhost:8080'
...
paths:
...
  /api/transactions/address/{address}:
    get:
      description: This request retrieves all transactions sorted by AVAX
      operationId: getTransactionsByAddress
      parameters:
        - in: query
          required: false
          name: direction
          description: |
            Transaction direction.\
            Sent=`OUT`, Received=`IN`, Both=`ALL`          
          schema:
            $ref: '#/components/schemas/DirectionParam'
        - in: path
          required: true
          name: address
          schema:
            type: string
      responses:
        200:
          description: Success          
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressTransaction'
        500:
          description: Internal Server Error          
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
      tags:
        - transactions
...
components:
  schemas:
...
    AddressTransaction:
      type: object
      properties:
        transactionId:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        transactionIndex:
          type: number
        blockNumber:
          type: number
        etherValue:
          type: number
        receiptStatus:
          type: string
          enum:
            - success
            - failed
        direction:
          type: string
          description: Sent=`OUT`, Received=`IN`
          enum: 
            - IN
            - OUT          
...
    DirectionParam:
      type: string
      description: |
        Transaction direction.\
        Sent=`OUT`, Received=`IN`, Both=`ALL`
      default: ALL      
      enum: 
        - IN
        - OUT
        - ALL
...
    ErrorInfo:
      type: object
      properties:
        exception:
          type: string
        message:
          type: string